dfsound - spu.c: silenced warning. - (win32) DFSound.rc, resource.h, cfg.c: Added mute option and freqresponse.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@61534 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfsound/spu.c
@@ -866,14 +866,14 @@ GOON: ;
 				tmp = -32768;
 			if (tmp > 32767)
 				tmp = 32767;
-			sl = tmp;
+			sl = (int)tmp;
 
 			tmp = avg + rdiff * _interpolation_coefficient;
 			if (tmp < -32768)
 				tmp = -32768;
 			if (tmp > 32767)
 				tmp = 32767;
-			sr = tmp;
+			sr = (int)tmp;
 
 
 			*pS++=sl/voldiv;
